refactor: move loadInstallSpec to shared file instead of separate package
Per feedback, creating a whole package for just LoadInstallSpec was
overkill. Move the function to a shared.go file within the cmd package
instead for better code organization without the overhead of a separate
package.
